chore: rough conversion to TypeScript
