+++ src/commands/apply.ts
@@ -1,9 +1,11 @@
+import djs from 'discord.js'
+
 module.exports = {
     name: 'apply',
     description: 'Sends the server application.',
     aliases: ['app'],
-    async execute(client, message, args, fart, config){
-        const pingEmbed = new fart.MessageEmbed()
+    async execute(client:djs.Client, message:djs.Message, args, config){
+        const pingEmbed : djs.MessageEmbed = new djs.MessageEmbed()
         .setColor(config["main_config"].colorhex)
         .setAuthor(`${message.guild.name}'s Server Application`)
         .setDescription(`[Click Me](${config["rp_bot_config"].applylink})`)
