+++ src/commands/ip.ts
@@ -1,9 +1,11 @@
+import djs from 'discord.js'
+
 module.exports = {
     name: 'ip',
     description: 'Sends the server ip.',
     aliases: ['ip2'],
-    async execute(client, message, args, fart, config){
-        const pingEmbed = new fart.MessageEmbed()
+    async execute(client:djs.Client, message:djs.Message, args, config){
+        const pingEmbed : djs.MessageEmbed= new djs.MessageEmbed()
         .setColor(config["main_config"].colorhex)
         .setAuthor(`${message.guild.name}'s Server IP`)
         .setDescription(`${config["rp_bot_config"].ip}`)
