+++ src/handlers/Event.ts
@@ -1,0 +1,14 @@
+import fs from 'fs'
+
+module.exports = (client, djs, config) => {
+    const load_dir = (dirs) => {
+        const event_files = fs.readdirSync(`@/events/${dirs}`).filter(file => file.endsWith(`.js`));
+        
+        for (const file of event_files) {
+            const event = require(`@/events/${dirs}/${file}`);
+            const event_name = file.split('.')[0];
+            client.on(event_name, event.bind(null, client, djs, config));
+        }
+    }
+    ['client', 'guild'].forEach(e => load_dir(e));
+}
